requirements.txt
fastapi
litellm
opencv-python
python-dotenv
numpy
uvicorn
python-multipart